build: silence third-party warnings, keep own-code warnings; fix size_t narrowings
- /utf-8 global (C too) to stop samplerate C4819 spam
- md_silence_external() silences samplerate/kaldi/zxing targets
- third-party includes as SYSTEM to stop header-warning leaks
- move /W3 and -Wall/-Wextra to ModelDeploySDK target (md_warn_own)
- fix C4267 narrowings in utils.cpp and C4018 sign-check in tensor.cpp

+++ csrc/core/tensor.cpp
@@ -705,7 +705,7 @@ namespace modeldeploy {
 
         size_t index = 0;
         for (size_t i = 0; i < indices.size(); ++i) {
-            if (indices[i] >= shape_[i]) {
+            if (indices[i] >= static_cast<size_t>(shape_[i])) {
                 throw std::out_of_range("Index out of range");
             }
             index += indices[i] * strides_[i];
